at least this is best version because I connected backend with frontend and even there was exceptions in code I fixed them, and I hope before tomorrow I will finish this project.

+++ src/main/java/org/example/library/model/Order.java
@@ -35,5 +35,5 @@ public class Order {
 
     private int quantity; // Количество (обычно 1, но вдруг купят 2)
 
-    private int totalPrice; // Сумма заказа на момент покупки
+    private Double totalPrice; // Сумма заказа на момент покупки
 }
